CeleryProject/requirements.txt
Django==5.2.4
Pillow==9.0.1
Pillow==11.3.0
reportlab==3.6.8
